Adding a Postcodes::LocationChecker service
This commit adds the LocationChecker service to the postcodes domain.

Wasn't sure whether this should just take the postcode data as an
argument or the api_client itself. The latter approach allows us to only call
the API client if the postcodes are not on the allowed list so have gone
with that.